requirements.txt
transformers=4.6.0
tqdm>=4.28.1
torch>=1.7.0
pytorch_crf>=0.7.0
scikit_learn>=0.20.2
